@@ -23,13 +23,13 @@
     ],
 
     "require": {
-        "php": "^7.1",
-        "jms/serializer": "^1.4",
+        "php": "^7.4",
+        "jms/serializer": "^3.5",
         "guzzlehttp/guzzle": "^6.2"
     },
 
     "require-dev": {
-        "phpunit/phpunit": "^6.0"
+        "phpunit/phpunit": "^8.5"
     },
 
     "autoload": {
