@@ -30,7 +30,7 @@ class NetvisorTest extends \PHPUnit\Framework\TestCase
     /**
      *
      */
-    public function setUp()
+    public function setUp(): void
     {
         $this->client = $this->getMockBuilder('GuzzleHttp\Client')
             ->disableOriginalConstructor()
