@@ -15,7 +15,7 @@ class CustomerTest extends XmlTestCase
      */
     private $customer;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
