@@ -4,6 +4,7 @@ namespace Xi\Netvisor;
 
 use JMS\Serializer\Serializer;
 use JMS\Serializer\SerializerBuilder;
+use Xi\Netvisor\Component\Validate;
 use Xi\Netvisor\Serializer\Naming\LowercaseNamingStrategy;
 
 class XmlTestCase extends \PHPUnit\Framework\TestCase
@@ -13,12 +14,18 @@ class XmlTestCase extends \PHPUnit\Framework\TestCase
      */
     private $serializer;
 
-    public function setUp()
+	/**
+	 * @var Validate
+	 */
+	private $validate;
+
+    public function setUp(): void
     {
         $builder = SerializerBuilder::create();
         $builder->setPropertyNamingStrategy(new LowercaseNamingStrategy());
 
         $this->serializer = $builder->build();
+	    $this->validate = new Validate();
     }
 
     /**
@@ -37,9 +44,25 @@ class XmlTestCase extends \PHPUnit\Framework\TestCase
      */
     public function assertXmlContainsTagWithValue($tag, $value, $xml)
     {
-        $this->assertContains(sprintf('<%s', $tag), $xml);
-        $this->assertContains(sprintf('><![CDATA[%s]]></%s>', $value, $tag), $xml); // TODO: Integers are not enclosed with CDATA.
+	    $this->assertStringContainsString(sprintf('<%s', $tag), $xml);
+
+	    if (is_int($value) || is_float($value))
+	    {
+		    $this->assertStringContainsString(sprintf('>%s</%s>', $value, $tag), $xml);
+		    return;
+	    }
+
+	    $this->assertStringContainsString(sprintf('><![CDATA[%s]]></%s>', $value, $tag), $xml);
     }
+
+	/**
+	 * @param string $tag
+	 * @param string $xml
+	 */
+	public function assertXmlDoesNotContainTag($tag, $xml)
+	{
+		$this->assertStringNotContainsString(sprintf('<%s', $tag), $xml);
+	}
 
     /**
      * @param string $tag
@@ -54,6 +77,11 @@ class XmlTestCase extends \PHPUnit\Framework\TestCase
             $attributeLine .= sprintf(' %s="%s"', $key, $value);
         }
 
-        $this->assertContains(sprintf('<%s%s>', $tag, $attributeLine), $xml);
+        $this->assertStringContainsString(sprintf('<%s%s>', $tag, $attributeLine), $xml);
     }
+
+	public function assertXmlIsValid($xml, $dtdPath)
+	{
+		$this->assertTrue($this->validate->isValid($xml, $dtdPath));
+	}
 }
